src/python-backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
torch==2.2.0+cu121
torchvision==0.17.0+cu121
diffusers==0.30.0
transformers==4.44.0
accelerate==0.34.0
safetensors==0.4.5
peft==0.13.0
pillow==10.4.0
numpy==1.26.4
# LoRA Training dependencies
bitsandbytes==0.43.0  # 8-bit Adam optimizer for memory efficiency
datasets==2.16.0      # Dataset utilities
